[refactoring] step2: 리뷰 반영 패키지 수정
