+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,0 +1,9 @@
+package blackjack.domain
+
+import blackjack.entity.CardNumber
+import blackjack.entity.CardShape
+
+data class Card(
+    val number: CardNumber,
+    val shape: CardShape
+)
