--- src/main/kotlin/blackjack/entity/Card.kt
@@ -1,6 +1,0 @@
-package blackjack.entity
-
-data class Card(
-    val number: CardNumber,
-    val shape: CardShape
-)
