+++ src/main/kotlin/blackjack/entity/Cards.kt
@@ -1,5 +1,7 @@
 package blackjack.entity
 
+import blackjack.domain.Card
+
 data class Cards(
     val cards: ArrayDeque<Card>
 ) : List<Card> by cards {
