+++ src/test/kotlin/blackjack/card/CardsTest.kt
@@ -1,6 +1,6 @@
 package blackjack.card
 
-import blackjack.entity.Card
+import blackjack.domain.Card
 import blackjack.entity.CardNumber
 import blackjack.entity.CardShape
 import blackjack.entity.Cards
